azure-static-app/api/requirements.txt
azure-functions
requests
python-dotenv
azure-cognitiveservices-vision-computervision